Working scale and draw

+++ QML-project/Parts/KalkWykresSPLC/main.cpp
@@ -16,8 +16,8 @@ long double QmlCanvas::downY;
 int QmlCanvas::width;
 int QmlCanvas::height;
 
-long double QmlCanvas::perOnePixelX; //dopisać funkcje odwrotne
-long double QmlCanvas::perOnePixelY; //dopisać funkcje odwrotne
+double QmlCanvas::perOnePixelX; //dopisać funkcje odwrotne
+double QmlCanvas::perOnePixelY; //dopisać funkcje odwrotne
 int main(int argc, char *argv[])
 {
     QApplication app(argc, argv);
